% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{\code{fars_map_state} Uses make_filename and fars_read functions from this package to gather the desired year of accident file and
 plots a point map for desired state.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a numeric value (or string numberic value) corresponding to the desired state.}

\item{year}{a numeric value (or string numberic value) corresponding to the desired year of accident
reports.}
}
\value{
A point map for indicated state and year.
}
\description{
\code{fars_map_state} Uses make_filename and fars_read functions from this package to gather the desired year of accident file and
 plots a point map for desired state.
}
\note{
An error will occur if a value is entered to state.num which can not be converted to an integer.

An error will occur if a value is entered to state.num which doens't correspond to an existing state.

An error will occur if a value is entered to year which can not be converted to an integer.

A message will be issued if no accidents exist for the entered state or year.
}
\examples{
\dontrun{
fars_map_state("1", "2014")
}

}
