% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read()}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a character string of the file(s) path or file(s) name only (if file(s) is within work directory)
you would like to load into your data frame.}
}
\value{
A tibble and prints the tibble.
}
\description{
\code{fars_read()} Reads csv files and loads them into a dataframe table.
}
\details{
File to be loaded must be in a csv format.
}
\note{
An error will occur stating the given file doesn't exist if the path is invalid.
}
\examples{
\dontrun{
fars_read("C:/local/documents/accident_2014.csv")
}
}
