% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a numberic vector (or string of numberic values) containing the desired years of accident files
to read.}
}
\value{
A tibble with the month and year columns from each file read into the function.
}
\description{
\code{fars_read_years} #'Reads multiple accident files into a dataframe table using make_filename and fars_read functions from this
 package but requires the targeted files to be in the working directy.
}
\note{
A warning will appear for each year entered that doesn't have a corresponding accident file.

An error will occur if a none numberical entry is made.
}
\examples{
\dontrun{
fars_read_years("2013, 2014, 2015")
}
}
