% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename()}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a numerical entry which will be converted to an integer.}
}
\value{
A string in the following format: "accident_%d.csv.bz2" with the user entered year as an integer.
}
\description{
\code{make_filename()} Creates a file name for accident file(s) for a given a calander year.
}
\note{
An error will occur if a non-numberic entry is entered as an input.
}
\examples{
\dontrun{
make_filename("2014")
}

}
